fix: Fix insufficient index storage space
- Updated the configuration to set the volatile index directory to the same path as the persistent index directory to avoid running out of storage space.
- Added checks to skip persistence and preparation steps if both directories are identical, improving efficiency and reducing unnecessary operations.
- Create a symbolic link in tmpfs to keep the application access index entry unchanged.

Log: Fix insufficient index storage space

+++ src/server/src/core/file_index_manager.cpp
@@ -291,6 +291,11 @@ bool file_index_manager::commit(index_status status) {
 void file_index_manager::persist_index() {
     std::error_code ec;
 
+    if (persistent_index_directory_ == volatile_index_directory_) {
+        spdlog::debug("Persistent index directory is the same as volatile index directory, skip persist");
+        return;
+    }
+
     std::filesystem::remove_all(persistent_index_directory_, ec);
     if (ec) {
         spdlog::error("Failed to remove persistent index directory: {}", ec.message());
@@ -476,6 +481,28 @@ void file_index_manager::try_refresh_reader(bool nrt) {
 void file_index_manager::prepare_index() {
     spdlog::info("Preparing index...");
     std::error_code ec;
+
+    std::string index_entry_point = std::string(g_get_user_runtime_dir()) + "/deepin-anything-server";
+
+    if (volatile_index_directory_ != index_entry_point) {
+        // create symbolic links for index entry point
+        if (std::filesystem::exists(index_entry_point)) {
+            std::filesystem::remove_all(index_entry_point);
+        }
+        std::filesystem::create_directory_symlink(volatile_index_directory_, index_entry_point, ec);
+        if (ec) {
+            spdlog::error("Failed to create symbolic link from {} to {}: {}, quit",
+                index_entry_point, volatile_index_directory_, ec.message());
+            exit(APP_QUIT_CODE);
+        }
+    }
+
+    if (persistent_index_directory_ == volatile_index_directory_) {
+        spdlog::info("Persistent index directory is the same as volatile index directory, skip prepare");
+        return;
+    }
+
+    ec.clear();
     if (!std::filesystem::exists(volatile_index_directory_, ec)) {
         if (ec) {
             spdlog::error("Failed to check volatile index directory: {}", ec.message());
